+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - HTTP server implementation with health, debug and metrics endpoints
+- Deduplication task and subcommand
 
 ## [0.4.1] - 2023-10-24