+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use pipeline test report summary build ids to set test suite/case ids
 
+### Removed
+
+- Unused projects model
+
 ## [0.5.0] - 2024-01-26
 
 This is quite a big release that introduces a lot of breaking changes.