+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.1] - 2024-02-04
 
 ### Added
 
@@ -129,7 +129,8 @@ Initial release.
 
 <!-- Links -->
 
-[Unreleased]: https://github.com/cluttrdev/gitlab-exporter/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/cluttrdev/gitlab-exporter/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/cluttrdev/gitlab-exporter/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/cluttrdev/gitlab-exporter/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/cluttrdev/gitlab-exporter/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/cluttrdev/gitlab-exporter/compare/v0.3.1...v0.4.0